Made edges change color when picked/selected (still doesn't work 100%.)
Added status field to bottom of control panel.

Made picking and translating work at the same time when in PICKING mode.
Removed comboBox that was used to switch between PICKING and TRANSLATING
mode.

When selecting a property in the propTree, matching elements will be
picked/selected in the graph.

When right-clicking a property in the propTree, a context menu pops up.
Right now that menu has one item, which will copy a CYPHER expression
matching the property's key value pair.